requirements.txt
# Summer School Chatbot Dependencies

# Core dependencies
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# LLM and AI
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Web framework (for future interface)
flask>=2.3.0
flask-cors>=4.0.0
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0 